requirements.txt
#Flask==0.12.2
pymessenger==0.0.7.0
gunicorn==19.6.0
tensorflow-cpu==2.6.0

Flask==1.1.2
#gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.18.5

